move testing page from main activity to TestScreen.kt and TestViewModel.kt. restructure ui files to be in ui/ instead of ui/theme.
