+++ app/src/main/java/com/example/recipeplanner/ui/CreateRecipeScreen.kt
@@ -1,0 +1,4 @@
+package com.example.recipeplanner.ui
+
+class CreateRecipeScreen {
+}
