+++ app/src/main/java/com/example/recipeplanner/ui/HomeScreenViewModel.kt
@@ -1,0 +1,8 @@
+package com.example.recipeplanner.ui
+
+import androidx.lifecycle.ViewModel
+
+
+class HomeScreenViewModel() : ViewModel() {
+
+}
