--- app/src/main/java/com/example/recipeplanner/ui/theme/CreateRecipeScreen.kt
@@ -1,4 +1,0 @@
-package com.example.recipeplanner.ui.theme
-
-class CreateRecipeScreen {
-}
