--- app/src/main/java/com/example/recipeplanner/ui/theme/CreateRecipeViewModel.kt
@@ -1,4 +1,0 @@
-package com.example.recipeplanner.ui.theme
-
-class CreateRecipeViewModel {
-}
